move meta classes to meta package
