+++ corehq/apps/api/domain_metadata.py
@@ -1,11 +1,11 @@
 import logging
 
 from corehq.apps.api.resources.auth import AdminAuthentication
+from corehq.apps.api.resources.meta import CustomResourceMeta
 from corehq.apps.api.serializers import XFormInstanceSerializer
 from corehq.apps.domain.models import Domain
 from corehq.apps.accounting.models import Subscription
 from corehq.apps.api.resources import HqBaseResource, CouchResourceMixin
-from corehq.apps.api.resources.v0_1 import CustomResourceMeta
 from corehq.apps.es.domains import DomainES
 
 from tastypie import fields
