+++ corehq/apps/locations/resources/v0_1.py
@@ -3,7 +3,7 @@ from tastypie import fields
 from tastypie.resources import Resource
 
 from corehq.apps.api.resources.auth import LoginAndDomainAuthentication
-from corehq.apps.api.resources.v0_1 import CustomResourceMeta
+from corehq.apps.api.resources.meta import CustomResourceMeta
 from corehq.apps.api.util import get_object_or_not_exist
 from corehq.apps.api.resources import HqBaseResource
 from corehq.apps.domain.models import Domain
